Revert "Issue #SB-17509 fix: Bootstrapping app only after splash screen is show"
